bots/.template/pyproject.toml
[tool.poetry]
name = "template_project"
version = "0.1.0"
description = ""
authors = ["Hannah Willoughby <hannah@hannahwilloughby.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
"Mastodon.py" = "^1.5.1"
pytest = "^7.1.2"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[tool.pytest.ini_options]
minversion = "7.1"
addopts = "-ra -q"
testpaths = [
    "."
]